+++ src/main/java/act/controller/bytecode/ControllerByteCodeScanner.java
@@ -21,9 +21,7 @@ package act.controller.bytecode;
  */
 
 import act.Act;
-import act.app.App;
-import act.app.AppByteCodeScannerBase;
-import act.app.AppClassLoader;
+import act.app.*;
 import act.app.event.SysEventId;
 import act.asm.*;
 import act.asm.signature.SignatureReader;
@@ -34,9 +32,7 @@ import act.controller.annotation.Port;
 import act.controller.annotation.TemplateContext;
 import act.controller.meta.*;
 import act.handler.builtin.controller.RequestHandlerProxy;
-import act.route.DuplicateRouteMappingException;
-import act.route.RouteSource;
-import act.route.Router;
+import act.route.*;
 import act.sys.Env;
 import act.sys.meta.EnvAnnotationVisitor;
 import act.util.*;
@@ -45,13 +41,11 @@ import org.osgl.$;
 import org.osgl.http.H;
 import org.osgl.mvc.annotation.With;
 import org.osgl.mvc.util.Binder;
-import org.osgl.util.C;
-import org.osgl.util.E;
-import org.osgl.util.ListBuilder;
-import org.osgl.util.S;
+import org.osgl.util.*;
 
 import java.lang.annotation.Annotation;
 import java.util.*;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * New controller scanner implementation
@@ -213,7 +207,7 @@ public class ControllerByteCodeScanner extends AppByteCodeScannerBase {
                              * Note we need to schedule route registration after all app code scanned because we need the
                              * parent context information be set on class meta info, which is done after controller scanning
                              */
-                            app().jobManager().on(SysEventId.APP_CODE_SCANNED, new RouteRegister(envMatches, C.list(H.Method.GET), strings, WsEndpoint.PSEUDO_METHOD, routers, classInfo, false, $.var(false)));
+                            app().jobManager().on(SysEventId.APP_CODE_SCANNED, "WsEndpointAnnotationVisitor:registerRoute - " + registerRouteTaskCounter.getAndIncrement(), new RouteRegister(envMatches, C.list(H.Method.GET), strings, WsEndpoint.PSEUDO_METHOD, routers, classInfo, false, $.var(false)));
 
                             super.visitEnd();
                         }
@@ -773,6 +767,7 @@ public class ControllerByteCodeScanner extends AppByteCodeScannerBase {
                     }
                 }
 
+
                 @Override
                 public void visitEnd() {
                     super.visitEnd();
@@ -792,7 +787,7 @@ public class ControllerByteCodeScanner extends AppByteCodeScannerBase {
                      * Note we need to schedule route registration after all app code scanned because we need the
                      * parent context information be set on class meta info, which is done after controller scanning
                      */
-                    app().jobManager().on(SysEventId.APP_CODE_SCANNED, new RouteRegister(envMatched, httpMethods, paths, methodName, routers, classInfo, classInfo.isAbstract() && !isStatic, isVirtual));
+                    app().jobManager().on(SysEventId.APP_CODE_SCANNED, "ActionAnnotationVisitor:registerRoute-" + registerRouteTaskCounter.getAndIncrement(), new RouteRegister(envMatched, httpMethods, paths, methodName, routers, classInfo, classInfo.isAbstract() && !isStatic, isVirtual));
                 }
 
             }
@@ -1025,4 +1020,7 @@ public class ControllerByteCodeScanner extends AppByteCodeScannerBase {
             }
         }
     }
+
+    private static AtomicInteger registerRouteTaskCounter = new AtomicInteger(0);
+
 }