+++ src/main/java/act/job/meta/JobMethodMetaInfo.java
@@ -36,8 +36,12 @@ import org.osgl.util.S;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicInteger;
 
 public class JobMethodMetaInfo extends LogSupportedDestroyableBase {
+
+    private static final AtomicInteger jobIdCounter = new AtomicInteger(0);
+
     private String id;
     private String name;
     private InvokeType invokeType;
@@ -49,7 +53,7 @@ public class JobMethodMetaInfo extends LogSupportedDestroyableBase {
     public JobMethodMetaInfo(final JobClassMetaInfo clsInfo, final List<String> paramTypes) {
         this.clsInfo = clsInfo;
         final App app = Act.app();
-        app.jobManager().on(SysEventId.DEPENDENCY_INJECTOR_PROVISIONED, new Runnable() {
+        app.jobManager().on(SysEventId.DEPENDENCY_INJECTOR_PROVISIONED, "JobMethodMetaInfo:init-" + jobIdCounter.getAndIncrement(), new Runnable() {
             @Override
             public void run() {
                 $.Var<Method> var = $.var();