+++ src/main/java/act/view/rythm/RythmTransformerScanner.java
@@ -103,7 +103,7 @@ public class RythmTransformerScanner extends AppByteCodeScannerBase {
         public void visitEnd() {
             super.visitEnd();
             if (found) {
-                app().jobManager().on(SysEventId.PRE_START, new Runnable() {
+                app().jobManager().on(SysEventId.PRE_START, "RythmTransformerScanner:registerTransformer:" + className, new Runnable() {
                     @Override
                     public void run() {
                         RythmView rythmView = (RythmView) Act.viewManager().view(RythmView.ID);